Change include? to has_key? to fix ERROR: TypeError
Just saw this in IRC after experiencing it myself on a fresh omnibus
install of chef with knife-rackspace:

```
<ziftytodd> Just installed Chef on my workstation. Trying to use knife
rackspace server list/create/etc, but all variations result in ERROR:
TypeError: wrong argument type Symbol (expected Module). Any ideas on
what might be wrong?
```

The process I used to do the install:

* standard omnibus install from opscode install instructions for ubuntu
* /opt/chef/embedded/bin/gem install knife-rackspace (root via sudo -i)

This small fix I did live on my installation solved my issues but I
don't know if it is a workaround or a solution.  If anyone can let me
know if this is a good fix or also works for them that would be awesome.