Add boot_volume_size option to -B in `knife rackspace server create`
When using the -B option with `knife rackspace server create`, the size
of the created volume is hardcoded to 100GB.

In order to specify the volume size, I added the boot_volume_size option
to fog's create_server call (merged in fog 1.35).
This commit leverages this new fog option and allows to use smaller
volumes (miminum 50GB) as well as larger ones.

This commit also clarifies the wording of one misleading error message.